feat: change in routing logic, Now after a successful email verification, the user is redirected to the project view
