+++ backend/api/src/main/kotlin/io/tolgee/controllers/PublicController.kt
@@ -16,6 +16,7 @@ import io.tolgee.exceptions.NotFoundException
 import io.tolgee.model.UserAccount
 import io.tolgee.openApiDocs.OpenApiHideFromPublicDocs
 import io.tolgee.security.authentication.JwtService
+import io.tolgee.security.authorization.BypassEmailVerification
 import io.tolgee.security.payload.JwtAuthenticationResponse
 import io.tolgee.security.ratelimit.RateLimited
 import io.tolgee.security.thirdParty.GithubOAuthDelegate
@@ -146,6 +147,7 @@ class PublicController(
     description = "It checks whether the code from email is valid",
   )
   @OpenApiHideFromPublicDocs
+  @BypassEmailVerification
   fun verifyEmail(
     @PathVariable("userId") @NotNull userId: Long,
     @PathVariable("code") @NotBlank code: String,
