+++ webapp/src/component/RootRouter.tsx
@@ -1,20 +1,21 @@
-import React, { FC } from 'react';
-import { Redirect, Route, Switch } from 'react-router-dom';
-import { GoogleReCaptchaProvider } from 'react-google-recaptcha-v3';
+import React, {FC} from 'react';
+import {Redirect, Route, Switch} from 'react-router-dom';
+import {GoogleReCaptchaProvider} from 'react-google-recaptcha-v3';
 
-import { LINKS } from 'tg.constants/links';
-import { ProjectsRouter } from 'tg.views/projects/ProjectsRouter';
-import { UserSettingsRouter } from 'tg.views/userSettings/UserSettingsRouter';
-import { OrganizationsRouter } from 'tg.views/organizations/OrganizationsRouter';
-import { useConfig } from 'tg.globalContext/helpers';
-import { AdministrationView } from 'tg.views/administration/AdministrationView';
+import {LINKS} from 'tg.constants/links';
+import {ProjectsRouter} from 'tg.views/projects/ProjectsRouter';
+import {UserSettingsRouter} from 'tg.views/userSettings/UserSettingsRouter';
+import {OrganizationsRouter} from 'tg.views/organizations/OrganizationsRouter';
+import {useConfig} from 'tg.globalContext/helpers';
+import {AdministrationView} from 'tg.views/administration/AdministrationView';
 
-import { PrivateRoute } from './common/PrivateRoute';
-import { OrganizationBillingRedirect } from './security/OrganizationBillingRedirect';
-import { RequirePreferredOrganization } from '../RequirePreferredOrganization';
-import { HelpMenu } from './HelpMenu';
-import { PublicOnlyRoute } from './common/PublicOnlyRoute';
-import { PreferredOrganizationRedirect } from './security/PreferredOrganizationRedirect';
+import {PrivateRoute} from './common/PrivateRoute';
+import {OrganizationBillingRedirect} from './security/OrganizationBillingRedirect';
+import {RequirePreferredOrganization} from '../RequirePreferredOrganization';
+import {HelpMenu} from './HelpMenu';
+import {PublicOnlyRoute} from './common/PublicOnlyRoute';
+import {PreferredOrganizationRedirect} from './security/PreferredOrganizationRedirect';
+import {EmailNotVerifiedView} from 'tg.component/EmailNotVerifiedView';
 
 const LoginRouter = React.lazy(
   () => import(/* webpackChunkName: "login" */ './security/Login/LoginRouter')
@@ -78,6 +79,9 @@ const RecaptchaProvider: FC = (props) => {
 
 export const RootRouter = () => (
   <Switch>
+    <PrivateRoute exact path={LINKS.EMAIL_NOT_VERIFIED.template}>
+      <EmailNotVerifiedView />
+    </PrivateRoute>
     <PrivateRoute exact path={LINKS.SLACK_CONNECT.template}>
       <SlackConnectView />
     </PrivateRoute>
