+++ webapp/src/service/http/ApiHttpService.tsx
@@ -1,12 +1,12 @@
-import { LINKS } from 'tg.constants/links';
-import { GlobalError } from 'tg.error/GlobalError';
+import {LINKS} from 'tg.constants/links';
+import {GlobalError} from 'tg.error/GlobalError';
 
-import { tokenService } from '../TokenService';
-import { getUtmCookie } from 'tg.fixtures/utmCookie';
-import { handleApiError } from './handleApiError';
-import { ApiError } from './ApiError';
-import { errorAction } from './errorAction';
-import { globalContext } from 'tg.globalContext/globalActions';
+import {tokenService} from '../TokenService';
+import {getUtmCookie} from 'tg.fixtures/utmCookie';
+import {handleApiError} from './handleApiError';
+import {ApiError} from './ApiError';
+import {errorAction} from './errorAction';
+import {globalContext} from 'tg.globalContext/globalActions';
 
 let requests: { [address: string]: number } = {};
 const detectLoop = (url) => {
@@ -59,8 +59,9 @@ export class ApiHttpService {
             if (r.status >= 400) {
               const responseData = await ApiHttpService.getResObject(r);
               const resultError = new ApiError('Api error', responseData);
+              const currentPath = location.pathname;
               resultError.setErrorHandler(() =>
-                handleApiError(r, responseData, init, options)
+                handleApiError(r, responseData, init, options, currentPath)
               );
               if (r.status === 400) {
                 errorAction(responseData.code);
