+++ webapp/src/views/projects/ProjectListView.tsx
@@ -1,26 +1,20 @@
-import { useState } from 'react';
-import { T, useTranslate } from '@tolgee/react';
+import {useState} from 'react';
+import {T, useTranslate} from '@tolgee/react';
 
-import { EmptyListMessage } from 'tg.component/common/EmptyListMessage';
-import { PaginatedHateoasList } from 'tg.component/common/list/PaginatedHateoasList';
-import { BaseView } from 'tg.component/layout/BaseView';
-import { BaseViewAddButton } from 'tg.component/layout/BaseViewAddButton';
-import { DashboardPage } from 'tg.component/layout/DashboardPage';
-import { LINKS } from 'tg.constants/links';
-import { useApiQuery } from 'tg.service/http/useQueryApi';
+import {EmptyListMessage} from 'tg.component/common/EmptyListMessage';
+import {PaginatedHateoasList} from 'tg.component/common/list/PaginatedHateoasList';
+import {BaseView} from 'tg.component/layout/BaseView';
+import {BaseViewAddButton} from 'tg.component/layout/BaseViewAddButton';
+import {DashboardPage} from 'tg.component/layout/DashboardPage';
+import {LINKS} from 'tg.constants/links';
+import {useApiQuery} from 'tg.service/http/useQueryApi';
 import DashboardProjectListItem from 'tg.views/projects/DashboardProjectListItem';
-import { Button, styled } from '@mui/material';
-import { Link } from 'react-router-dom';
-import {
-  useEmailAwaitingVerification,
-  useIsAdmin,
-  useIsEmailVerified,
-  usePreferredOrganization,
-} from 'tg.globalContext/helpers';
-import { OrganizationSwitch } from 'tg.component/organizationSwitch/OrganizationSwitch';
-import { Usage } from 'tg.component/billing/Usage';
-import { QuickStartHighlight } from 'tg.component/layout/QuickStartGuide/QuickStartHighlight';
-import { EmailVerificationPage } from 'tg.component/EmailVerificationPage';
+import {Button, styled} from '@mui/material';
+import {Link} from 'react-router-dom';
+import {useIsAdmin, usePreferredOrganization,} from 'tg.globalContext/helpers';
+import {OrganizationSwitch} from 'tg.component/organizationSwitch/OrganizationSwitch';
+import {Usage} from 'tg.component/billing/Usage';
+import {QuickStartHighlight} from 'tg.component/layout/QuickStartGuide/QuickStartHighlight';
 
 const StyledWrapper = styled('div')`
   display: flex;
@@ -36,7 +30,6 @@ export const ProjectListView = () => {
   const [page, setPage] = useState(0);
   const [search, setSearch] = useState('');
   const { preferredOrganization } = usePreferredOrganization();
-  const isEmailVerified = useIsEmailVerified();
 
   const listPermitted = useApiQuery({
     url: '/v2/organizations/{slug}/projects-with-stats',
@@ -78,8 +71,7 @@ export const ProjectListView = () => {
           maxWidth={1000}
           allCentered
           addComponent={
-            addAllowed &&
-            isEmailVerified && (
+            addAllowed && (
               <QuickStartHighlight itemKey="add_project">
                 <BaseViewAddButton
                   addLinkTo={LINKS.PROJECT_ADD.build()}
@@ -90,45 +82,37 @@ export const ProjectListView = () => {
           }
           addLabel={t('projects_add_button')}
           hideChildrenOnLoading={false}
-          navigation={
-            isEmailVerified
-              ? [
-                  [<OrganizationSwitch key={0} />],
-                  [t('projects_title'), LINKS.PROJECTS.build()],
-                ]
-              : undefined
-          }
+          navigation={[
+            [<OrganizationSwitch key={0} />],
+            [t('projects_title'), LINKS.PROJECTS.build()],
+          ]}
           navigationRight={<Usage />}
           loading={listPermitted.isFetching}
         >
-          {!isEmailVerified ? (
-            <EmailVerificationPage email={useEmailAwaitingVerification()} />
-          ) : (
-            <PaginatedHateoasList
-              wrapperComponentProps={{ className: 'listWrapper' }}
-              onPageChange={setPage}
-              loadable={listPermitted}
-              renderItem={(r) => <DashboardProjectListItem key={r.id} {...r} />}
-              emptyPlaceholder={
-                <EmptyListMessage
-                  loading={listPermitted.isFetching}
-                  hint={
-                    isOrganizationOwner ? (
-                      <Button
-                        component={Link}
-                        to={LINKS.PROJECT_ADD.build()}
-                        color="primary"
-                      >
-                        <T keyName="projects_empty_action" />
-                      </Button>
-                    ) : undefined
-                  }
-                >
-                  <T keyName="projects_empty" />
-                </EmptyListMessage>
-              }
-            />
-          )}
+          <PaginatedHateoasList
+            wrapperComponentProps={{ className: 'listWrapper' }}
+            onPageChange={setPage}
+            loadable={listPermitted}
+            renderItem={(r) => <DashboardProjectListItem key={r.id} {...r} />}
+            emptyPlaceholder={
+              <EmptyListMessage
+                loading={listPermitted.isFetching}
+                hint={
+                  isOrganizationOwner ? (
+                    <Button
+                      component={Link}
+                      to={LINKS.PROJECT_ADD.build()}
+                      color="primary"
+                    >
+                      <T keyName="projects_empty_action" />
+                    </Button>
+                  ) : undefined
+                }
+              >
+                <T keyName="projects_empty" />
+              </EmptyListMessage>
+            }
+          />
         </BaseView>
       </DashboardPage>
     </StyledWrapper>
